Move CodeLensesProviderKeys into lang-commons module
