+++ language-server/modules/langserver-commons/src/main/java/org/ballerinalang/langserver/commons/codelenses/CodeLensesProviderKeys.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ballerinalang.langserver.codelenses;
+package org.ballerinalang.langserver.commons.codelenses;
 
 import org.ballerinalang.langserver.commons.LSContext;
 import org.ballerinalang.util.diagnostic.Diagnostic;
